+++ crates/store/src/server/mod.rs
@@ -11,8 +11,8 @@ use miden_node_proto_build::{
     store_rpc_api_descriptor,
 };
 use miden_node_utils::panic::{CatchPanicLayer, catch_panic_layer_fn};
+use miden_node_utils::signer::BlockSigner;
 use miden_node_utils::tracing::grpc::grpc_trace_fn;
-use miden_protocol::block::BlockSigner;
 use tokio::net::TcpListener;
 use tokio::task::JoinSet;
 use tokio_stream::wrappers::TcpListenerStream;
@@ -54,12 +54,13 @@ impl Store {
         skip_all,
         err,
     )]
-    pub fn bootstrap<S: BlockSigner>(
+    pub async fn bootstrap<S: BlockSigner>(
         genesis: GenesisState<S>,
         data_directory: &Path,
     ) -> anyhow::Result<()> {
         let genesis = genesis
             .into_block()
+            .await
             .context("failed to convert genesis configuration into the genesis block")?;
 
         let data_directory =