+++ crates/validator/src/lib.rs
@@ -1,9 +1,11 @@
 mod block_validation;
 mod db;
 mod server;
+mod signers;
 mod tx_validation;
 
 pub use server::Validator;
+pub use signers::{KmsSigner, ValidatorSigner};
 
 // CONSTANTS
 // =================================================================================================